requirements.txt
pyyaml
python-dotenv
openai
requests
